+++ app/javascript/app.vue
@@ -1,40 +1,37 @@
 <template>
-  <div id="app">
-    <table>
-      <tbody>
-        <tr>
-          <th>ID</th>
-          <th>Content</th>
-        </tr>
-        <tr v-for="tweet in tweets" :key="tweet.id">
-          <td>{{ tweet.id }}</td>
-          <td>{{ tweet.content }}</td>
-        </tr>
-      </tbody>
-    </table>
+  <div>
+    <router-link to="/">Index</router-link>
+    <router-view></router-view>
   </div>
 </template>
 
 <script>
-import axios from 'axios';
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+
+import TweetsIndexVue from 'TweetsIndex.vue'
+import TweetDetailVue from './TweetDetail.vue'
+
+const router = new VueRouter({
+  routes: [
+    {
+      path: '/',
+      component: TweetsIndexVue
+    },
+    {
+      path: '/tweets/:id(\\d+)',
+      name: 'TweetDetail',
+      component: TweetDetailVue
+    }
+  ]
+})
+
+Vue.use(VueRouter)
 
 export default {
-  data: function () {
-    return {
-      tweets: []
-    }
-  },
-  mounted () {
-    axios
-      .get('/api/v1/tweets.json')
-      .then(response => (this.tweets = response.data))
-  }
+  router
 }
 </script>
 
 <style scoped>
-p {
-  font-size: 2em;
-  text-align: center;
-}
 </style>
